+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * New/Edit Holding Record: Move Save button to fixed footer component Refs UIIN-916.
 * New/Edit Item Record: Move Save button to fixed footer component. Refs UIIN-917.
 * Add missing location permissions to `ui-inventory.instance.view`. Fixes UIIN-1056.
+* Item. Remove The Edit pen icon in the top main pane. Fixes UIIN-1026.
 
 ## [2.0.0](https://github.com/folio-org/ui-inventory/tree/v2.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.1...v2.0.0)